+++ apps/ledger-live-desktop/static/i18n/en/app.json
@@ -5053,13 +5053,13 @@
     "screens": {
       "welcome": {
         "videos": {
-          "buySell": "Buy, sell, swap, stake, and manage your portfolio",
-          "thousandsCrypto": "Access thousands of crypto in one ecosystem",
-          "secureWallet": "Step into the most secure wallet, wherever you go"
+          "buySell": "A wallet that protects and puts you in control",
+          "thousandsCrypto": "Send, receive, swap and stake thousands of crypto",
+          "secureWallet": "Verify all your transactions with peace of mind"
         },
         "ledgerDevice": "I have a Ledger device",
         "noDeviceYet": "I don't, yet",
-        "notice": "By continuing, you agree to Ledger's <Link1>Terms and Conditions</Link1> and <Link2>Privacy Policy</Link2>. Ledger provides no financial advice.\nTransaction services are provided by third parties.",
+        "notice": "By continuing, you agree to Ledger's <Link1>Terms & Conditions</Link1> and <Link2>Privacy Policy</Link2>. Ledger provides no financial advice.\nTransaction services are provided by third parties.",
         "title": "The most secure cryptocurrency & NFT wallet",
         "description": "Securely manage, buy and grow your crypto and NFTs on Ledger Live.",
         "nextButton": "Get started",
